chore: move columns, add kube-util
Moves columns to kube/column/Name (new naming pattern) and moves the
isNamespaced() function to kube-util.ts for future sharing.
