+++ packages/renderer/src/lib/configmaps-secrets/ConfigMapSecretList.svelte
@@ -10,7 +10,7 @@ import {
 } from '/@/stores/kubernetes-contexts-state';
 
 import ConfigMapSecretIcon from '../images/ConfigMapSecretIcon.svelte';
-import KubernetesColumnName from '../objects/KubernetesColumnName.svelte';
+import NameColumn from '../kube/column/Name.svelte';
 import KubernetesObjectsList from '../objects/KubernetesObjectsList.svelte';
 import { ConfigMapSecretUtils } from './configmap-secret-utils';
 import ConfigMapSecretColumnActions from './ConfigMapSecretColumnActions.svelte';
@@ -36,7 +36,7 @@ let statusColumn = new TableColumn<ConfigMapSecretUI>('Status', {
 
 let nameColumn = new TableColumn<ConfigMapSecretUI>('Name', {
   width: '1.3fr',
-  renderer: KubernetesColumnName,
+  renderer: NameColumn,
   comparator: (a, b): number => a.name.localeCompare(b.name),
 });
