+++ packages/renderer/src/lib/kube/column/Name.spec.ts
@@ -21,8 +21,8 @@ import '@testing-library/jest-dom/vitest';
 import { fireEvent, render, screen } from '@testing-library/svelte';
 import { expect, test } from 'vitest';
 
-import KubernetesColumnName from './KubernetesColumnName.svelte';
-import type { KubernetesNamespacedObjectUI, KubernetesObjectUI } from './KubernetesObjectUI';
+import type { KubernetesNamespacedObjectUI, KubernetesObjectUI } from '../../objects/KubernetesObjectUI.ts';
+import KubernetesColumnName from './Name.svelte';
 
 const node: KubernetesObjectUI = {
   kind: 'Node',
