+++ packages/renderer/src/lib/kube/kube-util.spec.ts
@@ -1,0 +1,44 @@
+/**********************************************************************
+ * Copyright (C) 2025 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+
+import { expect, test } from 'vitest';
+
+import type { KubernetesNamespacedObjectUI, KubernetesObjectUI } from '../objects/KubernetesObjectUI';
+import { isNamespaced } from './kube-util';
+
+const node: KubernetesObjectUI = {
+  kind: 'Node',
+  name: 'my-node',
+  status: '',
+};
+
+const deployment: KubernetesNamespacedObjectUI = {
+  kind: 'Deployment',
+  name: 'my-deployment',
+  status: '',
+  namespace: 'default',
+  selected: false,
+};
+
+test('isNamespaced is false for nodes', async () => {
+  expect(isNamespaced(node)).toBeFalsy();
+});
+
+test('isNamespaced is true for deployments', async () => {
+  expect(isNamespaced(deployment)).toBeTruthy();
+});
