+++ packages/renderer/src/lib/service/ServicesList.svelte
@@ -5,7 +5,7 @@ import moment from 'moment';
 import { kubernetesCurrentContextServicesFiltered, serviceSearchPattern } from '/@/stores/kubernetes-contexts-state';
 
 import ServiceIcon from '../images/ServiceIcon.svelte';
-import KubernetesColumnName from '../objects/KubernetesColumnName.svelte';
+import NameColumn from '../kube/column/Name.svelte';
 import KubernetesObjectsList from '../objects/KubernetesObjectsList.svelte';
 import { ServiceUtils } from './service-utils';
 import ServiceColumnActions from './ServiceColumnActions.svelte';
@@ -35,7 +35,7 @@ let statusColumn = new TableColumn<ServiceUI>('Status', {
 
 let nameColumn = new TableColumn<ServiceUI>('Name', {
   width: '1.3fr',
-  renderer: KubernetesColumnName,
+  renderer: NameColumn,
   comparator: (a, b): number => a.name.localeCompare(b.name),
 });
